Issue and Merge Request events
The issue id and title are always presented. Links to the issue. If the
description was changed, that is mentioned, likewise if the title is changed.
If both are changed then only the description is mentioned. The description is
only presented upon opening the issue, or if the description is changed. If
the issue is updated after having been closed, it is mentioned as a 'closed'
issue, or 'merged' merge request. If commits are added to a merge request,
this is mentioned, but only if the merge request is coming from a fork.